HeroesResponse works

+++ app/src/main/java/com/example/tinkofflab2023/data/remote/response/matches/TowerDamage.kt
@@ -7,5 +7,5 @@ data class TowerDamage(
     @SerializedName("pct")
     val pct: Double,
     @SerializedName("raw")
-    val raw: Int
+    val raw: Double
 )
